Force the XDMA_EVENT_INTR0 to be a GPIO instead of a clock line.
This is being done in the i2c2 pin mux area because that is something we
specifically need for the E3B and I cannot figure out where that clock
line is being enabled to turn it off in the DTS file. I know that we can
disable it here, so that is what I am doing.